requirements.txt
asgiref==3.5.0
Django==4.0.2
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
filter==0.0.0.20200724
importlib-metadata==4.11.3
Markdown==3.3.7
phonenumbers==8.12.47
Pillow==9.0.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2021.5
zipp==3.8.0
